Check function for multiple values for an area in an indicator for spine chart

Check function for multiple values for an area in an indicator for spine chart

spine_data_check(data, indicator, area_code)

Arguments

data
a data frame to create the spine chart from. the data frame should contain data for all area types included in the chart (eg, if plotting for County & UA with a comparator of region and a median line for national, the data frame should contain all of these data)

indicator
unquoted field name of the field containing the indicator labels. Take care as errors will occur where indicator labels are the same but data exist for multiple sub-categories (for example, sex or age)

area_code
unquoted field name where area codes are stored (local_area_code, median_line_area_code and comparator_area_code, if using, should all exist in this field)